physics system now uses floating points

+++ player.h
@@ -6,11 +6,10 @@
 
 struct protagonist
 {
-    int x;
-    int y;
-    // current velocities
-    int dx;
-    int dy;
+    float x;
+    float y;
+    float dx;     // current velocities
+    float dy;
     int inventory[MAX_ITEMS]; // 0 -> no item, 1 -> have item, 2 -> used item
     int frame;
 } player;
